+++ src/services/connectApiManager.js
@@ -1,16 +1,15 @@
 'use strict'
 
-const { invokeGet } = require('@shopware-pwa/shopware-6-client')
 const path = require('path')
 
 const getLoginEndpoint = () => 'sgwebcheckout/login'
 const getLoginTokenEndpoint = () => '/store-api/sgwebcheckout/login/token'
 
 /**
- * @param {ApiteSW6Utility.SWApiInstance?} api
+ * @param {AxiosInstance} api
  * @returns {Promise<SGConnectAPI.LoginTokenResponse>}
  */
-const getLoginToken = async (api) => invokeGet({ address: getLoginTokenEndpoint() }, api).then(({ data }) => data)
+const getLoginToken = async (api) => api.get(getLoginTokenEndpoint())
 
 /**
  * @param {string} baseUrl - the base Url to append login controller, e.g. http://example.com