+++ src/services/logDecorator.js
@@ -1,62 +1,106 @@
 'use strict'
 
+const get = require('lodash.get')
+
 /**
  * @param {string?} extensionName
  */
 const logDecorator = (extensionName = '@apite-shopware6-helper') => {
   /**
-   * @param {ApiteSW6Utility.SWClientApiError|ApiteSW6Utility.SWEntityError|ApiteSW6Utility.SWError|Error} error
-   * @return {ApiteSW6Utility.SWEntityError[]|ApiteSW6Utility.SWError[]|string[]}
+   * @param {AxiosResponse|ClientApiError|ShopwareError|EntityError|Error} response
+   * @return {SGConnectAPI.BasicErrorLog|SGConnectAPI.ClientApiErrorLog|FormattedAxiosResponse}
    */
-  const extractErrorMessages = (error) => {
-    if (error.statusCode) {
-      // SWClientApiError
-      return error.messages
-    } else if (error.messageKey || error.status) {
-      // SWEntityError | SWError
-      return [error]
+  const decorateError = response => {
+    if (response.statusText) {
+      return {
+        extension: extensionName,
+        ...formatAxiosResponse(response)
+      }
+    } else if (response.statusCode || response.statusCode === 0) {
+      return {
+        extension: extensionName,
+        errors: response.messages.map(err => ({
+          ...err,
+          meta: JSON.stringify(err.meta),
+          trace: JSON.stringify(get(err, 'trace[0]') || {})
+        }))
+      }
+    } else if (response.messageKey) {
+      return {
+        ...decorateMessage(response.message),
+        messageKey: response.messageKey
+      }
     }
-    // Error
-    return [error.message]
+
+    return {
+      ...decorateMessage(response.message),
+      ...response.stack ? { stack: _cleanMessage(response.stack) } : {}
+    }
   }
 
   /**
-   * @param {ApiteSW6Utility.SWClientApiError|ApiteSW6Utility.SWEntityError|ApiteSW6Utility.SWError|Error} error
-   * @return {string|number}
+   * @param {?string} message
+   * @return {SGConnectAPI.BasicErrorLog}
    */
-  const extractErrorCode = (error) => {
-    if (error.statusCode) {
-      return error.statusCode
-    } else if (error.messageKey) {
-      return error.code
-    } else if (error.status) {
-      return Number(error.status)
+  const decorateMessage = message => ({
+    extension: extensionName,
+    message: _cleanMessage(message)
+  })
+
+  /**
+   * @param {AxiosResponse} err
+   * @returns {FormattedAxiosRequest & FormattedAxiosResponse}
+   */
+  const formatAxiosResponse = err => (
+    {
+      ...formatAxiosRequest(err.config),
+      status: String(err.status),
+      statusText: err.statusText,
+      message: get(err, 'data.message') || '',
+      errors: Object.entries(get(err, 'data.errors') || {}).map(([, item]) => item.message)
     }
-    return 500
+  )
+
+  /**
+   * @param {AxiosRequestConfig} request
+   * @returns {FormattedAxiosRequest}
+   */
+  const formatAxiosRequest = request => {
+    return {
+      call: request.method.toUpperCase() + ' ' + request.baseURL + request.url,
+      headers: {
+        ...request.headers['sw-context-token'] ? { 'sw-context-token': obfuscateString(request.headers['sw-context-token']) } : {},
+        ...request.headers['sw-access-key'] ? { 'sw-access-key': obfuscateString(request.headers['sw-access-key']) } : {},
+        'sw-language-id': request.headers['sw-language-id']
+      },
+      params: request.params
+    }
   }
 
   /**
-   * @param {ApiteSW6Utility.SWClientApiError|ApiteSW6Utility.SWEntityError|ApiteSW6Utility.SWError|Error} error
-   * @return {{extension: string, code: (string|number), messages: (ApiteSW6Utility.SWEntityError[]|ApiteSW6Utility.SWError[]|string[])}}
+   * @param {?string} message
+   * @return {?string}
+   * @private
    */
-  const decorateError = (error) => ({
-    extension: extensionName,
-    code: extractErrorCode(error),
-    messages: extractErrorMessages(error)
-  })
+  const _cleanMessage = (message) => {
+    if (!message) {
+      return undefined
+    }
+    // Remove newline characters
+    let parsedQuery = message.replace(/\\n/g, '')
+    // Remove extra spaces
+    parsedQuery = parsedQuery.replace(/\s+/g, ' ').trim()
+
+    return parsedQuery !== '' ? parsedQuery : undefined
+  }
 
   /**
-   * @param {string} message
-   * @param {number} code - arbitrary code to search by
-   * @return {{extension: string, messages: string[]}}
+   * @param {?string} message
+   * @return {string}
    */
-  const decorateMessage = (message, code = 15) => ({
-    extension: extensionName,
-    code,
-    messages: [message]
-  })
+  const obfuscateString = message => message ? '***' + message.slice(-7) : message
 
-  return { decorateError, decorateMessage }
+  return { decorateError, decorateMessage, formatAxiosResponse, formatAxiosRequest, obfuscateString }
 }
 
 module.exports = logDecorator