+++ src/typedefs/models/checkout/cart/price/ReferencePrice.ts
@@ -1,0 +1,6 @@
+export interface ReferencePrice {
+  price: number;
+  purchaseUnit: number;
+  referenceUnit: number|undefined;
+  unitName: string;
+}