+++ src/typedefs/models/checkout/customer/ShippingAddress.ts
@@ -1,0 +1,6 @@
+import { CustomerAddress } from "./CustomerAddress";
+
+/**
+ * @public
+ */
+export interface ShippingAddress extends CustomerAddress {}