+++ src/typedefs/models/checkout/promotion/PromotionDiscountPrice.ts
@@ -1,0 +1,13 @@
+import { PromotionDiscount } from "./PromotionDiscount";
+import { Currency } from "../../system/currency/Currency";
+
+/**
+ * @public
+ */
+export interface PromotionDiscountPrice {
+  currencyId: string;
+  discountId: string;
+  price: number;
+  promotionDiscount: PromotionDiscount;
+  currency: Currency;
+}