+++ src/typedefs/models/checkout/promotion/PromotionIndividualCode.ts
@@ -1,0 +1,11 @@
+import { Promotion } from "./Promotion";
+
+/**
+ * @public
+ */
+export interface PromotionIndividualCode {
+  promotionId: string;
+  code: string;
+  promotion: Promotion | null;
+  payload: string[] | null;
+}