+++ src/typedefs/models/content/property/PropertyGroup.ts
@@ -1,0 +1,19 @@
+import { PropertyGroupOption } from "./PropertyGroupOption";
+import { CustomField } from "../../common/CustomField";
+import { PropertyGroupOptionTranslation } from "./PropertyGroupOptionTranslation";
+
+/**
+ * @public
+ */
+export interface PropertyGroup {
+  name: string;
+  displayType: string;
+  sortingType: string;
+  description: string | null;
+  options: PropertyGroupOption[] | null;
+  translations: PropertyGroupOptionTranslation[] | null;
+  customFields: CustomField[];
+  translated: {
+    [key: string]: any;
+  };
+}