+++ src/typedefs/response/ProductResult.ts
@@ -1,0 +1,10 @@
+import { PropertyGroup } from "../models/content/property/PropertyGroup";
+import { Product } from "../models/content/product/Product";
+
+/**
+ * @beta
+ */
+export interface ProductResponse {
+  product: Product;
+  configurator: PropertyGroup[];
+}